Add a new tree-based allocator.
It's bad, very bad, but it can free at least. A side effect of it's
memory usage is that we now need way more than 200M to run the kernel
and boot an application. I'm thinking I'll just look at some simple
allocation algorithm later and use that, but for now we have this.

Also, we really need kmalloc_aligned to work properly.